refactor: Make extensions use the vudeo-extractor lib
