+++ src/de/hdfilme/build.gradle
@@ -11,6 +11,7 @@ ext {
 }
 
 dependencies {
+    implementation(project(':lib-vudeo-extractor'))
     implementation(project(':lib-mixdrop-extractor'))
     implementation(project(':lib-dood-extractor'))
 }
