+++ src/es/pelisplushd/build.gradle
@@ -10,6 +10,7 @@ ext {
 }
 
 dependencies {
+    implementation(project(':lib-vudeo-extractor'))
     implementation(project(':lib-uqload-extractor'))
     implementation(project(':lib-streamwish-extractor'))
     implementation(project(':lib-filemoon-extractor'))
