+++ src/fr/vostfree/build.gradle
@@ -10,6 +10,7 @@ ext {
 }
 
 dependencies {
+    implementation(project(':lib-vudeo-extractor'))
     implementation(project(':lib-dood-extractor'))
     implementation(project(':lib-mytv-extractor'))
     implementation(project(':lib-okru-extractor'))
