+++ src/tr/turkanime/build.gradle
@@ -5,6 +5,7 @@ plugins {
 }
 
 dependencies {
+    implementation(project(':lib-vudeo-extractor'))
     implementation(project(':lib-uqload-extractor'))
     implementation "dev.datlag.jsunpacker:jsunpacker:1.0.1"
     implementation(project(":lib-cryptoaes"))
